LPD-6742 Refactoring Common, Extractor and Validator modules in just one module
