--- modules/util/portal-tools-db-partition-virtual-instance-migration-common/build.gradle
@@ -1,3 +1,0 @@
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-}
