--- modules/util/portal-tools-db-partition-virtual-instance-migration-extractor/build.gradle
@@ -1,46 +1,0 @@
-dependencies {
-	compileInclude group: "com.fasterxml.jackson.core", name: "jackson-annotations", version: "2.16.1"
-	compileInclude group: "com.fasterxml.jackson.core", name: "jackson-core", version: "2.16.1"
-	compileInclude group: "com.fasterxml.jackson.core", name: "jackson-databind", version: "2.16.1"
-	compileInclude group: "com.mysql", name: "mysql-connector-j", version: "8.3.0"
-	compileInclude group: "commons-cli", name: "commons-cli", version: "1.4"
-	compileInclude project(":util:portal-tools-db-partition-virtual-instance-migration-common")
-
-	compileOnly group: "com.liferay", name: "com.liferay.petra.function", version: "5.5.0"
-	compileOnly group: "com.liferay", name: "com.liferay.petra.lang", version: "5.2.0"
-	compileOnly group: "com.liferay", name: "com.liferay.petra.string", version: "5.3.0"
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "org.mockito", name: "mockito-core", version: "4.5.1"
-}
-
-deploy {
-	from "scripts"
-}
-
-distZip {
-	filesMatching("**/*.jar") {
-		it.path = it.name
-	}
-	rename('(.*)-(.*)(.jar)', '$1$3')
-}
-
-distributions {
-	main {
-		contents {
-			into("/") {
-				from "scripts"
-				from "README.markdown"
-			}
-		}
-	}
-}
-
-liferayOSGi {
-	expandCompileInclude = true
-}
-
-startScripts {
-	onlyIf {
-		false
-	}
-}
