--- modules/util/portal-tools-db-partition-virtual-instance-migration-extractor/src/main/resources/com/liferay/portal/tools/db/partition/virtual/instance/migration/extractor/util/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
