--- modules/util/portal-tools-db-partition-virtual-instance-migration-validator/README.markdown
@@ -1,21 +1,0 @@
-# Database Partition Virtual Instance Migration Validator Tool
-This tool validates all the conditions for a successful virtual instance migration are met.
-
-## Requirements
-- MySQL
-- Database user with DDL privileges
-
-## Usage
-```
-./db_partition_virtual_instance_migration_extractor.sh -s sourceFilePath -t targeFilePath
-```
-
-Options:
-- `-h`, `--help` Display options.
-- `-s`, `--source-file <arg>` Source file generated by the Database Partition Virtual Instance Migration Extractor Tool.
-- `-t` , `--target-file <arg>` Target file generated by the Database Partition Virtual Instance Migration Extractor Tool.
-
-### Execution example
-```
-./db_partition_virtual_instance_migration_validator.sh -s "/bundle/tools/portal-tools-db-partition-virtual-instance-migration-extractor/source.json" -t "/bundle/tools/portal-tools-db-partition-virtual-instance-migration-extractor/target.json"
-```
