--- modules/util/portal-tools-db-partition-virtual-instance-migration-validator/source-formatter-suppressions.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0"?>
-
-<suppressions>
-	<checkstyle>
-		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/validator/util/Validator\.java" />
-	</checkstyle>
-</suppressions>
