--- modules/util/portal-tools-db-partition-virtual-instance-migration-validator/src/main/resources/com/liferay/portal/tools/db/partition/virtual/instance/migration/validator/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
