+++ modules/util/portal-tools-db-partition-virtual-instance-migration/source-formatter-suppressions.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+
+<suppressions>
+	<checkstyle>
+		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/util/DatabaseUtil\.java" />
+		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/util/Validator\.java" />
+		<suppress checks="CompanyIterationCheck" files="src/test/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/DBPartitionVirtualInstanceMigrationExtractionTest\.java" />
+		<suppress checks="CompanyIterationCheck" files="src/test/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/util/DatabaseMockupUtil\.java" />
+	</checkstyle>
+</suppressions>
