+++ modules/util/portal-tools-db-partition-virtual-instance-migration/src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/common/Release.java
