+++ modules/util/portal-tools-db-partition-virtual-instance-migration/src/main/resources/com/liferay/portal/tools/db/partition/virtual/instance/migration/packageinfo
