+++ modules/util/portal-tools-db-partition-virtual-instance-migration/src/test/resources/com/liferay/portal/tools/db/partition/virtual/instance/migration/dependencies/sourceFailure.data
